requirements.txt
psycopg2-binary==2.9.9
asyncio==3.4.3
pyyaml==6.0.1
huggingface-hub==0.19.4
gpt-index==0.6.27
llama-index==0.6.27
cerebras-cloud-sdk==1.29.0